Move parser_utils
